pyproject.toml
[tool.poetry]
name = "atomichabits"
version = "0.1.0"
description = ""
authors = ["Kseniia <127607406+Kseniia-Artemova@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.0"
python-telegram-bot = "^20.6"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.0"
coverage = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
